airflow/requirements.txt
apache-airflow-providers-google
pyarrow==10.0.1